fixes a bug in SampleSignal where only the left channel of a stereo sound was being played.
fixes a bug in SampleSignal where the correct channel was not being returned from getChannel.

fixes a problem with the project settings where I'd set it to a specific JDK instead of leaving it as workspace default (I should have just changed the workspace default of my windows Eclipse install).

doing a new release: 2.0.1