Remove mention that Mac only listens to Mic-In
I have confirmed that getLineIn() will obtain the device selected in System Preferences, not just the default system input.  I tested using Loopback to create a virtual audio cable that routes the audio from applications into a usable microphone input for recording, streaming, etc. and was curious if it would obtain that input, or the microphone built into my laptop.  I'm glad to say, it turns out that it was the system audio mic that I selected!