Fix crash when changing wavetable too quickly
Hansi Raber ( @kritzikratzi ) gets credit for isolating and patching this bug that could cause minim to crash if changing the wavetable too quickly. Now it makes a local copy of the table, to avoid the possibility of its size changing while being processed. Something I ended up causing to happen regularly in my XYscope library which is constantly changing the wavetable on every draw cycle.